feat: הוספת מערכת הרשאות גרנולרית
- הוספת בדיקות הרשאות לכל פריטי הסיידבר (view permission)
- הוספת פונקציות עזר ב-api-auth.php: requireViewPermission, requireCreatePermission, requireEditPermission, requireDeletePermission
- עדכון כל קבצי ה-API עם בדיקות הרשאות לכל פעולה (list, get, create, update, delete)
- משתמש ללא הרשאת צפייה לא יראה את הלשונית בסיידבר ויחסם גם ברמת ה-API

+++ dashboard/dashboards/cemeteries/api/residency-api.php
@@ -11,25 +11,30 @@ $action = $_GET['action'] ?? '';
 // טיפול בפעולות שונות
 switch ($action) {
     case 'list':
+        requireViewPermission('residency');
         listResidencies();
         break;
-        
+
     case 'get':
+        requireViewPermission('residency');
         getResidency($_GET['id'] ?? '');
         break;
-        
+
     case 'save':
+        requireEditPermission('residency');
         saveResidency();
         break;
-        
+
     case 'delete':
+        requireDeletePermission('residency');
         deleteResidency($_GET['id'] ?? '');
         break;
-        
+
     case 'search':
+        requireViewPermission('residency');
         searchResidencies($_GET['query'] ?? '');
         break;
-        
+
     default:
         echo json_encode(['success' => false, 'error' => 'Invalid action']);
 }
